requirements.txt
streamlit
yfinance
pandas
numpy
scipy
GoogleNews
google-generativeai